refactor: decouple core logic from global configuration
