+++ cmd/front/main.go
@@ -4,9 +4,9 @@
 package main
 
 import (
-	"github.com/sonr-io/motr/config"
 	"github.com/sonr-io/motr/middleware"
 	"github.com/sonr-io/motr/routes"
+	"github.com/sonr-io/motr/sink/config"
 )
 
 func main() {
