+++ middleware/context.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/labstack/echo/v4"
 	"github.com/segmentio/ksuid"
-	"github.com/sonr-io/motr/config"
 	"github.com/sonr-io/motr/sink"
+	"github.com/sonr-io/motr/sink/config"
 	"github.com/sonr-io/motr/sink/models"
 	"github.com/syumai/workers/cloudflare/kv"
 )
